Tests/LibWeb/Text/input/navigation/iframe-pushstate-before-nested-history-ready.html

<!doctype html>
<body></body>
<script src="../include.js"></script>
<script>
    asyncTest(async done => {
        const server = httpTestServer();
        const finalPageUrl = await server.createEcho("GET", "/iframe-pushstate-before-nested-history-ready-final", {
            status: 200,
            headers: { "Content-Type": "text/html" },
            body: `<!DOCTYPE html><script>parent.postMessage("final page loaded", "*");<\/script>`,
        });

        const iframe = document.createElement("iframe");

        document.body.append(iframe);
        iframe.contentWindow.history.pushState({}, "", "about:blank#child");

        await internals.flushSessionHistoryTraversalQueue();
        println(iframe.contentWindow.location.hash);
        println("queue drained");

        const loaded = new Promise(resolve => {
            window.addEventListener("message", event => {
                if (event.source === iframe.contentWindow && event.data === "final page loaded")
                    resolve();
            });
        });
        iframe.src = finalPageUrl;
        await loaded;
        println("navigation drained");
        done();
    });
</script>
