Tests/LibWeb/Text/expected/navigation/iframe-pushstate-before-nested-history-ready.txt

#child
queue drained
navigation drained
